+++ tests/e2e_memory_operations.rs
@@ -288,7 +288,7 @@ async fn test_concurrent_memory_operations() -> Result<()> {
                     content: format!("Concurrent memory {}", i),
                     embedding: None,
                     tier: Some(MemoryTier::Working),
-                    importance_score: Some(0.5 + (i as f32 * 0.05)),
+                    importance_score: Some(0.5 + (i as f64 * 0.05)),
                     metadata: Some(serde_json::json!({
                         "test_id": test_id,
                         "concurrent_test": true,
@@ -391,7 +391,7 @@ async fn test_performance_under_load() -> Result<()> {
                     content: format!("Performance test memory {} with some additional content to make it realistic", i),
                     embedding: None,
                     tier: Some(MemoryTier::Working),
-                    importance_score: Some(0.5 + (i as f32 % 10) * 0.05),
+                    importance_score: Some(0.5 + ((i % 10) as f64) * 0.05),
                     metadata: Some(serde_json::json!({
                         "test_id": test_id,
                         "performance_test": true,
